% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgamfac.R
\name{lgamfac}
\alias{lgamfac}
\title{Logistic Gamma Lasso factor analysis for Binomial data}
\usage{
lgamfac(X, max_n = NULL, numfac, cl = NULL, covars = NULL,
  numfac_init = NULL, override = FALSE, NUM_ITER = 5)
}
\arguments{
\item{X}{an integer matrix (presence-absence/genotype/ generic binomial success matrix)}

\item{max_n}{The integer/matrix recording the number of trials from which matrix X is created.
Defaults to an integer corresponding to the maximum integer value of X.}

\item{numfac}{The number of logistic factors (does not include the intercept).}

\item{cl}{The cluster process created before running the code.}

\item{covars}{A matrix of known covariates which is modeled together with with the unknown factros. 
Defaults to NULL.}

\item{numfac_init}{Initial number of factors in the pre-processing steps before gamma lasso analysis.
Chosen to be bigger than \code{\numfac}.}

\item{override}{optional boolean to bypass Lanczos bidiagonalization
SVD. Usually not advised unless encountering a bug in the SVD code.}

\item{NUM_ITER}{The number of iterative steps of gamma-lasso estimation of factors.}
}
\value{
matrix of logistic factors (F) and corresponding loadings (L).
}
\description{
Fit a sparse logistic factor analysis model equipped with  Gamma Lasso shrinkage
priors on binomial data.
}
\note{
The matrix \code{X} is a binomial matrix. If it is a matrix of presence - absence 
(0/1), the \code{max_n} would be 1, if it is a genotype matrix, 0, 1 or 2 coded, 
then \code{max_n} would be 2.  We can also have a general X matrix recording the 
number of successes at each cell, with \code{max_n} being a matrix of same dimension as X
and recording the total numbe rof trials at each cell.
}
